feat: START modularizing with utils.js showModal and refreshInterval
